dashboard/requirements.txt
google-cloud-firestore
google-cloud-storage
google-cloud-vision
meraki==0.80.3
opencv-python
pystache
pytz
requests
requests_toolbelt
